+++ cw-orch-interchain/examples/timeout_packet.rs
@@ -6,7 +6,7 @@ use cosmos_sdk_proto::{
     traits::{Message, Name},
     Any,
 };
-use cw_orch::{environment::QueryHandler, prelude::*, tokio::runtime::Runtime};
+use cw_orch::{environment::QueryHandler, prelude::*};
 use cw_orch_interchain_core::InterchainEnv;
 use cw_orch_interchain_daemon::ChannelCreator as _;
 use cw_orch_starship::Starship;
@@ -14,8 +14,7 @@ use ibc_relayer_types::core::ics24_host::identifier::PortId;
 fn main() -> cw_orch::anyhow::Result<()> {
     pretty_env_logger::init();
 
-    let runtime = Runtime::new()?;
-    let starship = Starship::new(runtime.handle(), None)?;
+    let starship = Starship::new(None)?;
     let interchain = starship.interchain_env();
 
     let channel = interchain.create_channel(
@@ -59,7 +58,7 @@ fn main() -> cw_orch::anyhow::Result<()> {
         None,
     )?;
 
-    let result = interchain.wait_ibc("juno-1", tx_resp)?;
+    let result = interchain.await_packets("juno-1", tx_resp)?;
 
     match &result.packets[0].outcome {
         cw_orch_interchain_core::types::IbcPacketOutcome::Timeout { .. } => {}