+++ cw-orch/examples/local_daemon.rs
@@ -1,11 +1,7 @@
 use counter_contract::{
-    msg::{ExecuteMsg, GetCountResponse, InstantiateMsg, QueryMsg},
-    CounterContract,
+    msg::InstantiateMsg, CounterContract, CounterExecuteMsgFns, CounterQueryMsgFns,
 };
-use cw_orch::{
-    environment::Environment,
-    prelude::{CwOrchExecute, CwOrchInstantiate, CwOrchQuery, CwOrchUpload, Daemon, TxHandler},
-};
+use cw_orch::{environment::Environment, prelude::*};
 
 // From https://github.com/CosmosContracts/juno/blob/32568dba828ff7783aea8cb5bb4b8b5832888255/docker/test-user.env#L2
 const LOCAL_MNEMONIC: &str = "clip hire initial neck maid actor venue client foam budget lock catalog sweet steak waste crater broccoli pipe steak sister coyote moment obvious choose";
@@ -39,9 +35,9 @@ pub fn main() {
     assert!(init_res.is_ok());
     // ANCHOR_END: daemon_usage
 
-    let exec_res = counter.execute(&ExecuteMsg::Increment {}, None);
+    let exec_res = counter.increment();
     assert!(exec_res.is_ok());
 
-    let query_res = counter.query::<GetCountResponse>(&QueryMsg::GetCount {});
+    let query_res = counter.get_count();
     assert!(query_res.is_ok());
 }