+++ CHANGELOG.md
@@ -7,6 +7,13 @@
 - Deprecate cw-orch-core `TxHandler::sender` in favor of `TxHandler::sender_addr`
 - Implement `SenderBuilder`, `QuerySender` and `TxSender` which allow for customizing the transaction commitment logic.
 - Can now easily build `QueryOnlyDaemon` which will only expose query functions.
+- Changed cw-orch-interchain InterchainEnv API
+  - `chain` --> `get_chain`
+  - `follow_packet` --> `await_single_packet`
+  - `wait_ibc` --> `await_packets`
+  - `check_ibc` --> `await_and_check_packets`
+  - `follow_packets_from_tx_hash` --> `await_packets_for_txhash`
+- Better Docs for interchain, cw-orch and clone-testing
 
 ### Breaking